enforce unique flash sales
